requirements.txt
torch>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
transformers>=4.15.0
datasets>=2.0.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
ViennaRNA>=2.5.0
esm>=1.0.0
networkx>=2.5
dgl>=1.0.0
scipy>=1.7.0
biopython>=1.79
huggingface_hub>=0.12.0
tqdm>=4.62.0
multimolecule>=0.1.0 